+++ Variant7.h
@@ -26,4 +26,4 @@ constexpr double Y_max = 1.0;
 constexpr double X_min = -X_max;
 constexpr double Y_min = -Y_max;
 
-void CreateMatrixesV7(CSRMatrix& A, std::vector<double>& f, __int64 M, __int64 N);
+void CreateMatrixesV7(CSRMatrix& A, std::vector<double>& f, int M, int N);
